fix bugs for lfuda logic

+++ src/rgw/rgw_directory.cc
@@ -601,7 +601,8 @@ int RGWBlockDirectory::getAvgCacheWeight(string endpoint){
   });
 
   client.sync_commit();
-  return stoull(val);
+  //return stoull(val);
+  return stoi(val);
 }
 /*
   int min = INT_MAX;
